+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/DefaultForeignKeyDefinition.java
@@ -87,23 +87,28 @@ final class DefaultForeignKeyDefinition extends AbstractAttributeDefinition<Enti
 		private final Set<Column<?>> readOnlyColumns = new HashSet<>(1);
 		private final EntityType referencedEntityType;
 		private final int fetchDepth;
-		private final boolean soft;
 
 		private List<Attribute<?>> attributes = emptyList();
+		private boolean soft = false;
 
-		DefaultForeignKeyDefinitionBuilder(ForeignKey foreignKey, int fetchDepth, boolean soft) {
+		DefaultForeignKeyDefinitionBuilder(ForeignKey foreignKey, int fetchDepth) {
 			super(foreignKey);
 			if (fetchDepth < -1) {
 				throw new IllegalArgumentException("Fetch depth must be at least -1: " + foreignKey);
 			}
 			this.referencedEntityType = foreignKey.referencedType();
-			this.soft = soft;
 			this.fetchDepth = fetchDepth;
 		}
 
 		@Override
 		public ForeignKeyDefinition build() {
 			return new DefaultForeignKeyDefinition(this);
+		}
+
+		@Override
+		public ForeignKeyDefinition.Builder soft(boolean soft) {
+			this.soft = soft;
+			return this;
 		}
 
 		@Override